requirement.txt
pycryptodome
maskpass
bcrypt
progress
cryptography
pyfiglet
colorama